Move function to DebuggerWidget file
